# spec/api-spec.yaml
openapi: 3.0.0
info:
  title: 'Distancia Backend Service'
  description: 'This service provides endpoints for all related interactions'
  version: 1.0.0
  contact:
    name: 'Distancia Team'
    email: 'dev@distancia.com'
...
servers:
  - description: Staging
    url: https://distancia-dtl.herokuapp.com/
...
security:
  - bearerToken: []
paths:
...
  /user/login:
    post:
      summary: 'Endpoint to login user'
      operationId: loginUser
      x-visibility: public
      tags:
        - 'user'

      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/loginUser'

      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        'default':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
...
    loginUser:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          description: The email of the user
        password:
          type: string
          description: The password of the user
...
    LoginUserResponse:
      type: 'object'
      required:
        - 'email'
        - 'username'
        - 'id'
        - 'token'
      properties:
        id:
          type: 'string'
          format: 'uuid'
          description: 'The id of the user'
        email:
          type: 'string'
          description: 'The user email'
        username:
          type: 'string'
          description: 'The username'
        token:
          type: string
          description: 'The token'
...
    ErrorMessageDto:
      $ref: './common.yaml#/components/schemas/ErrorMessageDto'
...
  responses:
...
    Unauthorized:
      description: Client is not authorized to make request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessageDto'
          example:
            code: 'UNAUTHORIZED'
            message: 'Either client security header is missing or it is not valid.'
...
    ServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessageDto'
          example:
            code: 'INTERNAL_ERROR'
            message: 'An internal server error occurred.'
  securitySchemes:
    bearerToken:
      type: 'http'
      scheme: 'bearer'
      bearerFormat: 'JWT'
      description: 'Authentication for user-facing endpoints. Uses bearer token in API prefix.'
    authToken:
      type: 'apiKey'
      name: 'x-auth-token'
      in: 'header'
      description: 'Auth Token header for inter-service communication'